requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.22
asyncpg==0.29.0
python-dotenv==1.0.0
psycopg-binary==3.3.3
aiohttp==3.9.1
openpyxl==3.1.2
aiosqlite==0.20.0
pytz==2025.2
